+++ mujoco_ros/include/mujoco_ros/viewer.h
@@ -182,6 +182,7 @@ public:
 	mjOption mjopt_prev_;
 	mjvOption opt_prev_;
 	mjvCamera cam_prev_;
+
 	int warn_vgeomfull_prev_;
 
 	// pending GUI-driven actions, to be applied at the next call to Sync
@@ -302,6 +303,7 @@ public:
 
 	// additional user-defined viszualization geoms
 	mjvScene *user_scn = nullptr;
+	mjtByte user_scn_flags_prev_[mjNRNDFLAG];
 
 	// OpenGL rendering and UI
 	int refresh_rate   = 60;